+++ Assets/Scripts/Entity/IEntityStatus.cs
@@ -3,7 +3,7 @@ public interface IEntityStatus {
     int Mana { get; }
     int Morale { get; }
     int AttackDamage { get; }
-    int AttackSpeed { get; }
+    float AttackSpeed { get; }
     int MoveSpeed { get; }
     int AttackRange { get; }
 }
